pyproject.toml
[project]
name = "aion-talk-scripts"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
    "pillow",
]
